Interface and type refactoring
